+++ exporters/otlp/otlp_integration_test.go
@@ -94,13 +94,13 @@ func newExporterEndToEndTest(t *testing.T, additionalOpts []otlp.ExporterOption)
 		),
 	}
 	tp1 := sdktrace.NewTracerProvider(append(pOpts,
-		sdktrace.WithResource(resource.New(
+		sdktrace.WithResource(resource.NewWithAttributes(
 			label.String("rk1", "rv11)"),
 			label.Int64("rk2", 5),
 		)))...)
 
 	tp2 := sdktrace.NewTracerProvider(append(pOpts,
-		sdktrace.WithResource(resource.New(
+		sdktrace.WithResource(resource.NewWithAttributes(
 			label.String("rk1", "rv12)"),
 			label.Float32("rk3", 6.5),
 		)))...)
@@ -492,6 +492,7 @@ func TestNewExporter_withMultipleAttributeTypes(t *testing.T) {
 			sdktrace.WithMaxExportBatchSize(10),
 		),
 	)
+	defer func() { _ = tp.Shutdown(context.Background()) }()
 
 	tr := tp.Tracer("test-tracer")
 	testKvs := []label.KeyValue{