+++ exporters/stdout/doc.go
@@ -14,4 +14,8 @@
 
 // Package stdout contains an OpenTelemetry exporter for both tracing and
 // metric telemetry to be written to an output destination as JSON.
+//
+// This package is currently in a pre-GA phase. Backwards incompatible changes
+// may be introduced in subsequent minor version releases as we work to track
+// the evolving OpenTelemetry specification and user feedback.
 package stdout // import "go.opentelemetry.io/otel/exporters/stdout"