+++ sdk/metric/controller/controllertest/test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package controllertest
+package controllertest // import "go.opentelemetry.io/otel/sdk/metric/controller/controllertest"
 
 import (
 	"time"