+++ sdk/metric/processor/reducer/reducer_test.go
@@ -75,9 +75,7 @@ func TestFilterProcessor(t *testing.T) {
 	)
 	accum := metricsdk.NewAccumulator(
 		reducer.New(testFilter{}, processorTest.Checkpointer(testProc)),
-		metricsdk.WithResource(
-			resource.New(label.String("R", "V")),
-		),
+		resource.NewWithAttributes(label.String("R", "V")),
 	)
 	generateData(accum)
 
@@ -94,9 +92,7 @@ func TestFilterBasicProcessor(t *testing.T) {
 	basicProc := basic.New(processorTest.AggregatorSelector(), export.CumulativeExporter)
 	accum := metricsdk.NewAccumulator(
 		reducer.New(testFilter{}, basicProc),
-		metricsdk.WithResource(
-			resource.New(label.String("R", "V")),
-		),
+		resource.NewWithAttributes(label.String("R", "V")),
 	)
 	exporter := processorTest.NewExporter(basicProc, label.DefaultEncoder())