+++ sdk/trace/provider.go
@@ -12,13 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package trace
+package trace // import "go.opentelemetry.io/otel/sdk/trace"
 
 import (
+	"context"
 	"sync"
 	"sync/atomic"
 
-	"go.opentelemetry.io/otel"
+	"go.opentelemetry.io/otel/global"
+	"go.opentelemetry.io/otel/trace"
+
 	export "go.opentelemetry.io/otel/sdk/export/trace"
 	"go.opentelemetry.io/otel/sdk/instrumentation"
 	"go.opentelemetry.io/otel/sdk/resource"
@@ -46,7 +49,7 @@ type TracerProvider struct {
 	config         atomic.Value // access atomically
 }
 
-var _ otel.TracerProvider = &TracerProvider{}
+var _ trace.TracerProvider = &TracerProvider{}
 
 // NewTracerProvider creates an instance of trace provider. Optional
 // parameter configures the provider with common options applicable
@@ -80,8 +83,8 @@ func NewTracerProvider(opts ...TracerProviderOption) *TracerProvider {
 
 // Tracer with the given name. If a tracer for the given name does not exist,
 // it is created first. If the name is empty, DefaultTracerName is used.
-func (p *TracerProvider) Tracer(name string, opts ...otel.TracerOption) otel.Tracer {
-	c := otel.NewTracerConfig(opts...)
+func (p *TracerProvider) Tracer(name string, opts ...trace.TracerOption) trace.Tracer {
+	c := trace.NewTracerConfig(opts...)
 
 	p.mu.Lock()
 	defer p.mu.Unlock()
@@ -141,7 +144,7 @@ func (p *TracerProvider) UnregisterSpanProcessor(s SpanProcessor) {
 	}
 	if stopOnce != nil {
 		stopOnce.state.Do(func() {
-			s.Shutdown()
+			global.Handle(s.Shutdown(context.Background()))
 		})
 	}
 	if len(new) > 1 {
@@ -178,6 +181,21 @@ func (p *TracerProvider) ApplyConfig(cfg Config) {
 		c.Resource = cfg.Resource
 	}
 	p.config.Store(&c)
+}
+
+// Shutdown shuts down the span processors in the order they were registered
+func (p *TracerProvider) Shutdown(ctx context.Context) error {
+	spss, ok := p.spanProcessors.Load().(spanProcessorStates)
+	if !ok || len(spss) == 0 {
+		return nil
+	}
+
+	for _, sps := range spss {
+		sps.state.Do(func() {
+			global.Handle(sps.sp.Shutdown(ctx))
+		})
+	}
+	return nil
 }
 
 // WithSyncer registers the exporter with the TracerProvider using a